Refactor: useNav 커스텀 훅 작성
