+++ src/components/Nav/Nav.js
@@ -6,26 +6,21 @@ import ExtendedSearchBar from './ExtendedSearchBar/ExtendedSearchBar';
 import Login from '../../pages/Login/Login';
 import { useNavigate } from 'react-router-dom';
 import { Link } from 'react-router-dom';
+import useNav from './useNav';
 
 const Nav = () => {
-  const [isMenuExtend, setIsMenuExtend] = useState(false);
-  const [isSearchExtend, setIsSearchExtend] = useState(false);
-  const [isModalOpen, setIsModalOpen] = useState(false);
-  const navigate = useNavigate();
-
-  const modalClose = () => {
-    setIsModalOpen(!isModalOpen);
-  };
-  const token = localStorage.getItem('token') || '';
-
-  useEffect(() => {
-    setIsModalOpen(false);
-  }, [token]);
-
-  const onLogout = () => {
-    localStorage.removeItem('token');
-    alert('로그아웃이 되었습니다.');
-  };
+  const {
+    isMenuExtend,
+    setIsMenuExtend,
+    isSearchExtend,
+    setIsSearchExtend,
+    navigate,
+    modalClose,
+    onLogout,
+    token,
+    isModalOpen,
+    setIsModalOpen,
+  } = useNav();
 
   return (
     <>
